moved test into lazy-render-test.js
